// web/src/api/generated/model/sessionInfo.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Baba.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export interface SessionInfo {
  accountsOn: boolean;
  signedIn: boolean;
  /** @nullable */
  userName: string | null;
  /** @nullable */
  displayName: string | null;
  /** @nullable */
  roleNameEn: string | null;
  /** @nullable */
  roleNameAr: string | null;
  mustChangePassword: boolean;
  permissions: string[];
}